+++ src/WPBT/WPBT_Core.php
@@ -401,6 +401,7 @@ class WPBT_Core {
 
 		// Set base version for development channel if necessary.
 		$current_exploded = array_map( 'intval', explode( '.', $exploded_version[0] ) );
+
 		if ( 'development' === self::$options['channel'] && isset( $current_exploded[2] ) ) {
 			$current_exploded[1] = ++$current_exploded[1];
 			unset( $current_exploded[2] );