+++ src/WPBT/WP_Beta_Tester.php
@@ -228,7 +228,7 @@ class WP_Beta_Tester {
 	 * Get current WP release version.
 	 *
 	 * @since 3.1.0
-	 * @return string $wp_version
+	 * @return string|array $wp_version
 	 */
 	public function get_current_wp_release() {
 		$response = get_site_transient( 'current_wp_release' );
@@ -238,7 +238,7 @@ class WP_Beta_Tester {
 			$response = wp_remote_retrieve_body( $response );
 
 			if ( is_wp_error( $response ) ) {
-				return null;
+				return array();
 			}
 
 			$response = (array) json_decode( $response );