+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* return empty array for 8.1 compatibility
 
 #### 3.2.3 / 2022-09-29
 * update for PHP 8.1 compatibility